# docs/architecture/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baseline",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/checkins/daily": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Submit Daily Checkin",
        "operationId": "submit_daily_checkin_v1_checkins_daily_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyCheckInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_DailyCheckInResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analysis/daily": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Generate Daily Analysis",
        "operationId": "generate_daily_analysis_v1_analysis_daily_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_DailyAnalysisResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/briefings/{date}": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "Get Daily Briefing",
        "operationId": "get_daily_briefing_v1_briefings__date__get",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_DailyBriefingResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assistant/query": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Ask Assistant",
        "operationId": "ask_assistant_v1_assistant_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_AssistantQueryResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recommendations/{id}/feedback": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Submit Recommendation Feedback",
        "operationId": "submit_recommendation_feedback_v1_recommendations__id__feedback_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_RecommendationFeedbackResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/data/export": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Export Data",
        "operationId": "export_data_v1_data_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_DataExportResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "APIEnvelope_AssistantQueryResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistantQueryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/APIResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIEnvelope[AssistantQueryResponse]"
      },
      "APIEnvelope_DailyAnalysisResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DailyAnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/APIResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIEnvelope[DailyAnalysisResponse]"
      },
      "APIEnvelope_DailyBriefingResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DailyBriefingResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/APIResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIEnvelope[DailyBriefingResponse]"
      },
      "APIEnvelope_DailyCheckInResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DailyCheckInResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/APIResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIEnvelope[DailyCheckInResponse]"
      },
      "APIEnvelope_DataExportResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataExportResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/APIResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIEnvelope[DataExportResponse]"
      },
...
      "APIEnvelope_RecommendationFeedbackResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationFeedbackResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/APIResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIEnvelope[RecommendationFeedbackResponse]"
      },
      "APIError": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "APIError"
      },
      "APIResponseMeta": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "v1"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "APIResponseMeta"
      },
...
      "AssistantQueryRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "question": {
            "type": "string",
            "minLength": 1,
            "title": "Question"
          },
          "date_context": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Context"
          },
          "allowed_data_scope": {
            "items": {
              "$ref": "#/components/schemas/DataScope"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed Data Scope"
          },
          "include_external_knowledge": {
            "type": "boolean",
            "title": "Include External Knowledge",
            "default": false
          },
          "privacy_mode": {
            "$ref": "#/components/schemas/PrivacyMode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "allowed_data_scope",
          "privacy_mode"
        ],
        "title": "AssistantQueryRequest"
      },
      "AssistantQueryResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "answer": {
            "type": "string",
            "minLength": 1,
            "title": "Answer"
          },
          "personal_evidence": {
            "items": {
              "$ref": "#/components/schemas/PersonalEvidence"
            },
            "type": "array",
            "minItems": 1,
            "title": "Personal Evidence"
          },
          "external_sources": {
            "items": {
              "$ref": "#/components/schemas/ExternalCitation"
            },
            "type": "array",
            "title": "External Sources"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceLevel"
          },
          "uncertainty": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Uncertainty"
          },
          "safety_status": {
            "$ref": "#/components/schemas/SafetyStatus"
          },
          "trace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "personal_evidence",
          "confidence",
          "uncertainty",
          "safety_status",
          "trace_id"
        ],
        "title": "AssistantQueryResponse"
      },
...
      "DailyAnalysisRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "force_recompute": {
            "type": "boolean",
            "title": "Force Recompute",
            "default": false
          },
          "include_external_knowledge": {
            "type": "boolean",
            "title": "Include External Knowledge",
            "default": false
          },
          "privacy_mode": {
            "$ref": "#/components/schemas/PrivacyMode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "privacy_mode"
        ],
        "title": "DailyAnalysisRequest"
      },
      "DailyAnalysisResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "analysis_job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisJobStatus"
          },
          "estimated_completion_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estimated Completion Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_job_id",
          "status",
          "estimated_completion_seconds"
        ],
        "title": "DailyAnalysisResponse"
      },
      "DailyBriefingResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "readiness_state": {
            "$ref": "#/components/schemas/ReadinessState"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceLevel"
          },
          "data_freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/PersonalEvidence"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence"
          },
          "recommendation_band": {
            "$ref": "#/components/schemas/RecommendationBand"
          },
          "candidate_options": {
            "items": {
              "$ref": "#/components/schemas/CandidateOption"
            },
            "type": "array",
            "title": "Candidate Options"
          },
          "goal_tradeoffs": {
            "items": {
              "$ref": "#/components/schemas/GoalTradeoff"
            },
            "type": "array",
            "title": "Goal Tradeoffs"
          },
          "uncertainty": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Uncertainty"
          },
          "safety_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Safety Notes"
          },
          "trace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trace Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "readiness_state",
          "confidence",
          "data_freshness",
          "evidence",
          "recommendation_band",
          "uncertainty",
          "safety_notes",
          "trace_id",
          "generated_at"
        ],
        "title": "DailyBriefingResponse"
      },
      "DailyCheckInFlags": {
        "properties": {
          "alcohol": {
            "type": "boolean",
            "title": "Alcohol",
            "default": false
          },
          "illness": {
            "type": "boolean",
            "title": "Illness",
            "default": false
          },
          "injury": {
            "type": "boolean",
            "title": "Injury",
            "default": false
          },
          "travel": {
            "type": "boolean",
            "title": "Travel",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DailyCheckInFlags"
      },
      "DailyCheckInRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "energy_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Score"
          },
          "mood_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood Score"
          },
          "soreness_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Soreness Score"
          },
          "stress_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stress Score"
          },
          "perceived_recovery_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Perceived Recovery Score"
          },
          "food_quality_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Quality Score"
          },
          "flags": {
            "$ref": "#/components/schemas/DailyCheckInFlags"
          },
          "structured_notes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Structured Notes"
          },
          "free_text_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text Note"
          },
          "sensitive_note_policy": {
            "$ref": "#/components/schemas/SensitiveNotePolicy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "sensitive_note_policy"
        ],
        "title": "DailyCheckInRequest"
      },
      "DailyCheckInResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "checkin_id": {
            "type": "string",
            "format": "uuid",
            "title": "Checkin Id"
          },
          "accepted_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Fields"
          },
          "redaction_status": {
            "$ref": "#/components/schemas/RedactionStatus"
          },
          "analysis_job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Job Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "checkin_id",
          "accepted_fields",
          "redaction_status"
        ],
        "title": "DailyCheckInResponse"
      },
      "DataExportFormat": {
        "type": "string",
        "enum": [
          "json",
          "csv"
        ],
        "title": "DataExportFormat"
      },
      "DataExportRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "export_scope": {
            "$ref": "#/components/schemas/DataExportScope"
          },
          "format": {
            "$ref": "#/components/schemas/DataExportFormat"
          },
          "include_raw_data": {
            "type": "boolean",
            "title": "Include Raw Data",
            "default": false
          },
          "include_model_traces": {
            "type": "boolean",
            "title": "Include Model Traces",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "export_scope",
          "format"
        ],
        "title": "DataExportRequest"
      },
      "DataExportResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "export_job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Export Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/DataExportStatus"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "export_job_id",
          "status",
          "expires_at"
        ],
        "title": "DataExportResponse"
      },
      "DataExportScope": {
        "type": "string",
        "enum": [
          "all",
          "health",
          "checkins",
          "briefings",
          "recommendations"
        ],
        "title": "DataExportScope"
      },
...
      "DataScope": {
        "type": "string",
        "enum": [
          "briefing_trace",
          "recent_health",
          "checkins",
          "memory",
          "goals",
          "external_knowledge"
        ],
        "title": "DataScope"
      },
...
      "FeedbackActionTaken": {
        "type": "string",
        "enum": [
          "followed",
          "partially_followed",
          "ignored",
          "planned"
        ],
        "title": "FeedbackActionTaken"
      },
      "FeedbackRating": {
        "type": "string",
        "enum": [
          "useful",
          "somewhat_useful",
          "not_useful",
          "unsafe_or_wrong"
        ],
        "title": "FeedbackRating"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrivacyMode": {
        "type": "string",
        "enum": [
          "local_only",
          "cloud_assisted",
          "hybrid"
        ],
        "title": "PrivacyMode"
      },
...
      "RecommendationFeedbackRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "rating": {
            "$ref": "#/components/schemas/FeedbackRating"
          },
          "action_taken": {
            "$ref": "#/components/schemas/FeedbackActionTaken"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "outcome_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rating",
          "action_taken"
        ],
        "title": "RecommendationFeedbackRequest"
      },
      "RecommendationFeedbackResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "feedback_id": {
            "type": "string",
            "format": "uuid",
            "title": "Feedback Id"
          },
          "memory_update_status": {
            "$ref": "#/components/schemas/MemoryUpdateStatus"
          },
          "eval_queue_status": {
            "$ref": "#/components/schemas/EvalQueueStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "feedback_id",
          "memory_update_status",
          "eval_queue_status"
        ],
        "title": "RecommendationFeedbackResponse"
      },
      "RedactionStatus": {
        "type": "string",
        "enum": [
          "redacted",
          "partial",
          "none"
        ],
        "title": "RedactionStatus"
      },
...
      "SensitiveNotePolicy": {
        "type": "string",
        "enum": [
          "exclude_from_external_llm",
          "summarize_before_external_llm",
          "allow_external_llm"
        ],
        "title": "SensitiveNotePolicy"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}